feat: rebuild api routes
